refactor: 커스텀 훅 명과 파일명이 일치하도록 수정
